requirements.txt
fastapi
SQLAlchemy
uvicorn[standard]==0.28.0
passlib
bcrypt==4.0.1
python-decouple
python-multipart
python-jose[cryptography]
psycopg2-binary
pymysql
alembic
pytest
pytest-asyncio
httpx
aiofiles
jinja2
pydantic
starlette
